Changing card to be a string, adding tests and some protection to addMoney func

+++ ATM_Project/DataBase.cpp
@@ -2,11 +2,15 @@
 //#include <filesystem>
 #include <iostream>
 
+#ifndef QT_NO_DEBUG
+#define IS_TRUE(x) { if (!x) std::cout << __FUNCTION__ << " failed on line " << __LINE__ << std::endl << std::flush; else std::cout << "ok\n"; }
+#endif
+
 DataBase::DataBase(std::string name)
 {
     //std::cout << QCoreApplication::applicationDirPath().toStdString() << '\n';
-    //std::string path = "/Users/akreidun/Desktop/MOOP_ATM/ATM_Project/";
-    std::string path = "../ATM_Project/";
+    std::string path = "/Users/akreidun/Desktop/MOOP_ATM/ATM_Project/";
+    //std::string path = "../ATM_Project/";
     std::string extention(".sqlite");
     std::string full_name = path + name + extention;
 
@@ -31,6 +35,14 @@ DataBase::DataBase(std::string name)
         //this->addCortege("1234123412341240", "1234", 0);
         //this->addCortege("1234123412341241", "4321", 11231.12321);
     }
+#ifndef QT_NO_DEBUG
+    std::cout << "********* Start testing of cardExist *********" << std::flush;
+    cardExistsTest();
+    std::cout << "********* Finished testing of cardExist *********" << std::flush;
+    std::cout << "********* Start testing of addMoneyTest *********" << std::flush;
+    addMoneyTest();
+    std::cout << "********* Finished testing of addMoneyTest *********" << std::flush;
+#endif
 }
 
 DataBase::~DataBase()
@@ -71,15 +83,19 @@ bool DataBase::addCortege(const std::string cardNo, const std::string pin, const
 
 bool DataBase::deleteCortege(const std::string cardNumber)
 {
+    if (!cardExists(cardNumber)) {
+        qDebug() << "There is no card with this number";
+        return false;
+    }
     QSqlQuery qry;
     std::string query("DELETE FROM card WHERE cardNo='" + cardNumber + "';");
     qry.prepare(query.c_str());
 
     if (!qry.exec()) {
         qDebug() << "error: deleting data from a database";
-        return false; // cortege was deleted
+        return false; // cortege wasn't deleted
     }
-    return true; // cortege wasn't deleted
+    return true; // cortege was deleted
 }
 
 bool DataBase::cardExists(const std::string cardNumber)
@@ -92,8 +108,22 @@ bool DataBase::cardExists(const std::string cardNumber)
         qDebug() << "error: getting data from a database";
         return false; // there is some error while executing query
     }
-
     return qry.next();
+}
+
+void DataBase::cardExistsTest()
+{
+    IS_TRUE(cardExists("1234123412341234"));
+    IS_TRUE(!cardExists("12341234123412344"));
+    IS_TRUE(cardExists("1234123412341235"));
+    IS_TRUE(cardExists("1234123412341238"));
+    IS_TRUE(cardExists("1234123412341239"));
+    IS_TRUE(cardExists("1234123412341240"));
+    IS_TRUE(cardExists("1234123412341241"));
+    IS_TRUE(!cardExists("1234123412341250"));
+    IS_TRUE(!cardExists("123123ed123"));
+    IS_TRUE(!cardExists("12321asfasf"));
+    IS_TRUE(!cardExists(""));
 }
 
 bool DataBase::checkTries(const std::string cardNumber)
@@ -190,13 +220,17 @@ double DataBase::getMoney(const std::string cardNumber)
 
 bool DataBase::addMoney(const std::string cardNumber, const double amount)
 {
+    if (amount < 0 && getMoney(cardNumber) < -amount) {
+        qDebug() << "There is not enough money on this card!";
+        return false; // adding was not performed
+    }
     if (!cardExists(cardNumber)) {
         qDebug() << "There is no card with this number";
         return false; // adding was not performed
     }
     double balance = getMoney(cardNumber);
     QSqlQuery qry;
-    std::string query("UPDATE card SET balance = '" + QString("%1").arg(balance + amount).toStdString() + "' WHERE cardNo='" + cardNumber + "';");
+    std::string query("UPDATE card SET balance = '" + QString("%3").arg(balance + amount).toStdString() + "' WHERE cardNo='" + cardNumber + "';");
     qry.prepare(query.c_str());
 
     if (!qry.exec()) {
@@ -204,6 +238,15 @@ bool DataBase::addMoney(const std::string cardNumber, const double amount)
         return false; // there is some error while executing query
     }
     return true; // Money added
+}
+
+void DataBase::addMoneyTest()
+{
+    IS_TRUE(!addMoney("12321", 1231));
+    IS_TRUE(!addMoney("12321", -1231));
+    IS_TRUE(addMoney("1234123412341234", 1231));
+    IS_TRUE(addMoney("1234123412341234", -1231));
+    IS_TRUE(!addMoney("1234123412341234", -1000000));
 }
 
 bool DataBase::getAllData()
@@ -219,16 +262,10 @@ bool DataBase::getAllData()
 
     bool check = false;
     while (qry.next()) {
-
-    // Here we need to save all Card objects according to this data
-
         QString cardNo = qry.value(0).toString();
         QString pin = qry.value(1).toString();
         double balance = qry.value(2).toDouble();
-
-    // ---
-
-        // qDebug() << cardNo << pin << balance;
+        qDebug() << cardNo << pin << balance;
         check = true;
     }
     if (!check) {
